% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarityFunc.R
\name{clustSimilarity}
\alias{clustSimilarity}
\title{clustSimilarity: main function to compare the similarity between genotypes
reported by souporcell and our known genotypes}
\usage{
clustSimilarity(
  pathToVCF,
  genotypes,
  IDs = NULL,
  compareIncomplete = FALSE,
  minQual = NULL,
  countReverseComplement = TRUE
)
}
\arguments{
\item{pathToVCF}{A string specifying a path to a VCF file produced by
souporcell. Provide full path including the file name.}

\item{genotypes}{A data.frame or string. String: current string option
is only "DO_mm10" (default) - automatically retrieves genotypes of 821
DO mice profiled in Farber lab. / Data.frame: You can provide your
genotype information. The first three columns are: marker (marker name),
chr (chromosome including "chr" prefix), position (marker position on a
chromosome). Following columns are assigned IDs. Numerical IDs must have
"X" prefix.}

\item{IDs}{Optional vector of strings specifying knwon IDs from genotype
data.frame. If IDs are numerical use "X" prefix in front of each ID.}

\item{compareIncomplete}{Logical variable specifying if only positions with
complete genotype information from known genotypes and souporcell output
should be used to calculate similarity. TRUE = compare markers for which
genotypes are NOT available for all tested individuals, FALSE =
compare only markers with genotypes available across all individials -
both from souporcell and known genotypes (default - FALSE, which means
only complete cases will be considered)}

\item{minQual}{An optional numerical variable specifying minimal QUAL
threshold for genotypes from souporcell vcf output to be considered}

\item{countReverseComplement}{Logical variable - an optional parameter
specifying if reverse complements should be considered matches.
Default: TRUE.}
}
\value{
data.frame containing similarities (proportional matches between
genotypes) between souporcell clusters (rows) and known genotypes
(columns)
}
\description{
clustSimilarity: main function to compare the similarity between genotypes
reported by souporcell and our known genotypes
}
\examples{
VCFpath = system.file("extdata", "cluster_genotypes.vcf",
                            package = "assignClustID")
genotypes = DO_mm10
similarities = clustSimilarity(pathToVCF=VCFpath, genotypes=genotypes)
}
