% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comparisonTableNoRevComp}
\alias{comparisonTableNoRevComp}
\title{Matrix containing dinucleotide matches excluding reverse complements}
\format{
A matrix with 16 rows and 16 variables: matches =1
}
\usage{
comparisonTableNoRevComp
}
\description{
Matrix containing dinucleotide matches excluding reverse complements
}
\keyword{datasets}
