% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DO_mm10}
\alias{DO_mm10}
\title{Dataset containing genotypes of selected DO mice}
\format{
A data.frame with 112913 rows and 9 variables
\describe{
\item{marker}{GigaMUGA markers}
\item{chr}{chromosome}
\item{position}{position of the marker on chromosome: mm10 reference}
\item{Xnum}{X followed by ID of a genotyped DO mouse}
}
}
\source{
\url{sftp://rivanna.hpc.virginia.edu/sfs/qumulo/qproject/farber_lab/Users/Current_users/Kristyna_K/annotations/DO_mice}
}
\usage{
DO_mm10
}
\description{
Dataset containing genotypes of selected DO mice
}
\keyword{datasets}
