% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarityFunc.R
\name{plotClustSimilarity}
\alias{plotClustSimilarity}
\title{plotClustSimilarity: plot results from \code{clustSimilarity} function in
form of heatmap}
\usage{
plotClustSimilarity(simMatrix, RCBpalette = "Purples")
}
\arguments{
\item{simMatrix}{Output from \code{clustSimilarity} function}

\item{RCBpalette}{Optional string specifying an RColorBrewer paltette to use
(default: "Purples")}
}
\value{
ComplexHeatmap object
}
\description{
plotClustSimilarity: plot results from \code{clustSimilarity} function in
form of heatmap
}
\examples{
VCFpath = system.file("extdata", "cluster_genotypes.vcf",
                            package = "assignClustID")
genotypes = DO_mm10
similarities = clustSimilarity(pathToVCF=VCFpath, genotypes=genotypes)
plotClustSimilarity(similarities)
}
