% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comparisonTable}
\alias{comparisonTable}
\title{Matrix containing all possible combinations of what could be dinucleotide
matches}
\format{
A matrix with 16 rows and 16 variables
\describe{
\item{dinucleotide}
}
}
\usage{
comparisonTable
}
\description{
Matrix containing all possible combinations of what could be dinucleotide
matches
}
\keyword{datasets}
